# Eio_linux: handle sleeping tasks that are now due before paused fibers