When using Eio I encountered an issue while porting a test in the mirage `arp` library. 

The failing code can be reduced to the following:
```ocaml
let rec loop () =
  Eio.Fiber.yield ();
  loop ()

let () = run @@ fun ~clock ->
  Fiber.first
    loop
    (fun () -> Eio.Time.sleep clock 0.1)
```

I expected this code to finish after `0.1s` of execution, but instead it does an infinite loop. This is because the scheduler prioritizes paused fibers before sleeping fibers that are now due. 

This PR adds a `Zzz.pop` before checking for paused fibers.