@TheLortex 2022-05-02
Thank you for your review.

> But with this change, make test_luv now hangs, so we probably need a similar fix there.

I have added a fix to the luv backend in the second commit. In particular I'm using `Luv.Async.send` to yield back to the event loop when the `IO` task is due.

> I think it would be simpler to delay adding the IO event until just before returning. 

Indeed that sounds better and reduces the diff. I have force-pushed my first commit with the suggested changes

> It still seems to be working, but might need to make it more robust, or at least update the comment to warn that it might not work now.

Do you have an idea on how to make it more robust ? Do you have tests that are waiting for the run queue to become empty ?