@TheLortex 2022-05-06
Thank you for performing the benchmarks. I expected a performance cost but not that high for the luv backend. A single task yielding is indeed the worst case, the run queue having 1 task and and 1 `IO` at any time, so the `IO` is performed once every two iterations. 

To address the performance cost globally, what we can do is remove the `IO` event from the run queue, and have instead a counter-based strategy:
- initially counter is at zero
- each run queue pop increases the counter
- when the counter is higher than a threshold (which value?), take the IO path instead of popping the next task
- when the IO path is taken set the counter to zero

This strategy has a negligible runtime cost when no busy yielding is happening. 
When busy yielding is happening, the cost will depend on on the number of tasks run between each IO lookup. The higher the threshold, the lower the overhead but it increases latency.

What do you think ?