@TheLortex 2022-05-17
I have no idea why the CI failure is happening. I have also been able to reproduce the test using the Docker instructions of the CI job..